Use state and parameters directly in DualConnection
