+++ src/main/java/com/atlassian/db/replica/internal/ReplicaConnectionProvider.java
@@ -1,42 +1,25 @@
 package com.atlassian.db.replica.internal;
 
 import com.atlassian.db.replica.api.reason.Reason;
-import com.atlassian.db.replica.api.reason.RouteDecision;
-import com.atlassian.db.replica.internal.logs.LazyLogger;
 import com.atlassian.db.replica.internal.state.ConnectionState;
-import com.atlassian.db.replica.internal.state.State;
-import com.atlassian.db.replica.spi.ReplicaConsistency;
 
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.SQLWarning;
-import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.Executor;
 
-import static com.atlassian.db.replica.api.reason.Reason.RO_API_CALL;
-import static com.atlassian.db.replica.internal.state.State.CLOSED;
-import static com.atlassian.db.replica.internal.state.State.MAIN;
 
-public class ReplicaConnectionProvider implements AutoCloseable {
-    private final ReplicaConsistency consistency;
+public class ReplicaConnectionProvider {
     private final ConnectionState state;
     private final ConnectionParameters parameters;
     private final Warnings warnings;
-    private final LazyLogger logger;
 
     public ReplicaConnectionProvider(
-        ReplicaConsistency consistency,
-        LazyLogger logger,
-        ConnectionParameters parameters,
-        Warnings warnings,
-        ConnectionState state
+        ConnectionParameters parameters, Warnings warnings, ConnectionState state
     ) {
         this.parameters = parameters;
         this.warnings = warnings;
         this.state = state;
-        this.consistency = consistency;
-        this.logger = logger;
     }
 
     public void setTransactionIsolation(Integer transactionIsolation) throws SQLException {
@@ -49,69 +32,6 @@ public class ReplicaConnectionProvider implements AutoCloseable {
         } else {
             return state.getWriteConnection(new RouteDecisionBuilder(Reason.RW_API_CALL)).getTransactionIsolation();
         }
-    }
-
-    public void setAutoCommit(Boolean autoCommit) throws SQLException {
-        final boolean autoCommitBefore = getAutoCommit();
-        if (autoCommitBefore != autoCommit) {
-            preCommit(autoCommitBefore);
-        }
-        parameters.setAutoCommit(state::getConnection, autoCommit);
-        if (autoCommitBefore != getAutoCommit()) {
-            recordCommit(autoCommitBefore);
-        }
-    }
-
-    public void setSchema(String schema) throws SQLException {
-        parameters.setSchema(state::getConnection, schema);
-    }
-
-    public void setClientInfo(ClientInfo clientInfo) throws SQLException {
-        parameters.setClientInfo(state::getConnection, clientInfo);
-    }
-
-    public void setNetworkTimeout(Executor executor, int milliseconds) throws SQLException {
-        parameters.setNetworkTimeout(state::getConnection, new NetworkTimeout(executor, milliseconds));
-    }
-
-    public boolean getAutoCommit() {
-        return parameters.isAutoCommit();
-    }
-
-    public Boolean isClosed() {
-        return state.getState().equals(CLOSED);
-    }
-
-    public boolean getReadOnly() {
-        return parameters.isReadOnly();
-    }
-
-    public void setReadOnly(boolean readOnly) throws SQLException {
-        parameters.setReadOnly(state::getConnection, readOnly);
-    }
-
-    public String getCatalog() {
-        return parameters.getCatalog();
-    }
-
-    public void setCatalog(String catalog) throws SQLException {
-        parameters.setCatalog(state::getConnection, catalog);
-    }
-
-    public Map<String, Class<?>> getTypeMap() {
-        return parameters.getTypeMap();
-    }
-
-    public void setTypeMap(Map<String, Class<?>> typeMap) throws SQLException {
-        parameters.setTypeMap(state::getConnection, typeMap);
-    }
-
-    public Integer getHoldability() throws SQLException {
-        return parameters.getHoldability() == null ? state.getWriteConnection(new RouteDecisionBuilder(Reason.RW_API_CALL)).getHoldability() : parameters.getHoldability();
-    }
-
-    public void setHoldability(Integer holdability) throws SQLException {
-        parameters.setHoldability(state::getConnection, holdability);
     }
 
     public SQLWarning getWarning() throws SQLException {
@@ -128,69 +48,5 @@ public class ReplicaConnectionProvider implements AutoCloseable {
             connection.get().clearWarnings();
         }
         warnings.clear();
-    }
-
-    public <T> T unwrap(Class<T> iface) throws SQLException {
-        final Connection currentConnection = state.getReadConnection(new RouteDecisionBuilder(RO_API_CALL));
-        if (iface.isAssignableFrom(currentConnection.getClass())) {
-            return iface.cast(currentConnection);
-        } else {
-            return currentConnection.unwrap(iface);
-        }
-    }
-
-    public boolean isWrapperFor(Class<?> iface) throws SQLException {
-        final Connection currentConnection = state.getReadConnection(new RouteDecisionBuilder(RO_API_CALL));
-        if (iface.isAssignableFrom(currentConnection.getClass())) {
-            return true;
-        } else {
-            return currentConnection.isWrapperFor(iface);
-        }
-    }
-
-    public State getState() {
-        return this.state.getState();
-    }
-
-    public void rollback() throws SQLException {
-        state.clearDirty();
-        final Optional<Connection> connection = state.getConnection();
-        if (connection.isPresent()) {
-            connection.get().rollback();
-        }
-    }
-
-    public void commit() throws SQLException {
-        final Optional<Connection> connection = state.getConnection();
-        if (connection.isPresent()) {
-            preCommit(parameters.isAutoCommit());
-            logger.debug(() -> "commit()");
-            connection.get().commit();
-            recordCommit(parameters.isAutoCommit());
-        }
-    }
-
-    private void recordCommit(boolean autoCommit) throws SQLException {
-        if (state.getState().equals(MAIN) && !autoCommit) {
-            final Connection mainConnection = state.getWriteConnection(new RouteDecisionBuilder(Reason.RW_API_CALL));
-            consistency.write(mainConnection);
-            state.clearDirty();
-        }
-    }
-
-    private void preCommit(boolean autoCommit) throws SQLException {
-        if (state.getState().equals(MAIN) && !autoCommit) {
-            final Connection mainConnection = state.getWriteConnection(new RouteDecisionBuilder(Reason.RW_API_CALL));
-            consistency.preCommit(mainConnection);
-        }
-    }
-
-    @Override
-    public void close() throws SQLException {
-        state.close();
-    }
-
-    public void abort(Executor executor) throws SQLException {
-        state.abort(executor);
     }
 }
